+++ designer/client/src/components/modals/TestingDataRecords/drawText.ts
