+++ designer/client/src/components/toolbars/test/buttons/ScenarioTestButton.tsx
@@ -16,7 +16,7 @@ import { useAppDispatch, useAppSelector } from "../../../../store/storeHelpers";
 import { useWindows, WindowKind } from "../../../../windowManager";
 import { getHasPendingChanges } from "../../../graph/node-modal/node/useEditState";
 import { useAdhocTestingAvailability } from "../../../modals/AdhocTesting/useAdhocTestingAvailability";
-import type { TestingData, TestingViewParams } from "../../../modals/Testing/TestingDialog";
+import type { TestingData, TestingViewParams } from "../../../modals/TestingDataRecords/Dialog";
 import { ButtonsVariant, ToolbarButton, ToolbarButtonsContext } from "../../../toolbarComponents/toolbarButtons";
 import { ToolbarSideContext } from "../../../toolbarComponents/ToolbarsContainer";
 import type { CustomButtonTypes, PropsOfButton } from "../../../toolbarSettings/buttons";
