+++ designer/client/src/http/HttpService/HttpService.ts
@@ -16,7 +16,7 @@ import type { ExpressionSuggestion } from "../../components/graph/node-modal/edi
 import type { AdditionalInfo } from "../../components/graph/node-modal/NodeAdditionalInfoBox";
 import { extractStickyNotesFromNodes } from "../../components/graph/utils/stickyNotesUtils";
 import type { AvailableScenarioLabels, ScenarioLabelsValidationResponse } from "../../components/Labels/types";
-import type { TestingEventParameters, TestingEventParametersRequestData } from "../../components/modals/Testing/TestingEventsTable";
+import type { TestingEventParameters, TestingEventParametersRequestData } from "../../components/modals/TestingDataRecords/Table";
 import type { ProcessName, ProcessVersionId, Scenario, StatusDefinitionType } from "../../components/Process/types";
 import type { ActivitiesResponse, ActivityMetadataResponse, ActivityType } from "../../components/toolbars/activities/types";
 import { ActivityTypesRelatedToExecutions } from "../../components/toolbars/activities/types";
